requirements.txt
docutils==0.15.2
pytz==2019.3
imutils==0.5.4
click==6.7
Jinja2==2.10.1
cryptography==3.2
numpy==1.18.1
itsdangerous==0.24
keyring==21.1.1
zipp==2.2.0
importlib_metadata==1.5.0
pytesseract==0.3.7
pywin32==227
Cython==0.29.15
Flask==1.1.1
Flask_PyMongo==2.3.0
Flask-API==1.0
Werkzeug==0.16.0
brotli==1.0.9
lockfile==0.12.2
ordereddict==1.1
protobuf==3.17.3
pyOpenSSL==20.0.1
python_bcrypt==0.3.2
scikit_learn==0.24.2
typing_extensions==3.10.0.0
